requirements.txt
flask
pdfminer.six
nltk
openai
python-dotenv
gunicorn